+++ ceammc/ext/src/base/xfade_tilde.h
@@ -26,13 +26,15 @@ enum XFadeFlags {
 
 class XFadeTilde : public SoundExternal {
 protected:
-    typedef SmoothLinT<t_float> t_smooth;
+    using t_smooth = SmoothLinT<t_float>;
     std::vector<t_smooth> gain_;
     t_float smooth_ms_;
     SymbolEnumProperty* prop_type_;
+    FloatProperty* value_;
 
 public:
     XFadeTilde(const PdArgs& args);
+    void initDone() override;
 
     void setupDSP(t_signal** in) override;
     void processBlock(const t_sample** in, t_sample** out) override;
@@ -42,6 +44,9 @@ public:
 public:
     // test
     std::vector<float> gains() const;
+
+private:
+    void setXFade(t_float v);
 };
 
 void setup_base_xfade_tilde();